se agregan autenticacion autorizacion jwt y guards de rutas

+++ src/mensajes/mensajes.service.ts
@@ -1,14 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Mensaje } from './entities/mensaje/mensaje.entity';
-import { Mensaje as mns } from 'src/interfaces/mensaje.interface';
 import { CreateMensajeDto } from './dto/create-mensaje-dto/create-mensaje-dto';
-
-
+import { UpdateMensajeDto } from './dto/create-mensaje-dto/update-mensajes-dto';
 @Injectable()
 export class MensajesService {
-    constructor(
+  constructor(
     @InjectRepository(Mensaje)
     private mensajesRepository: Repository<Mensaje>,
   ) {}
@@ -18,16 +16,18 @@ export class MensajesService {
     nuevo.mensaje = MensajeNuevo.mensaje;
     nuevo.nick = MensajeNuevo.nick;
     return this.mensajesRepository.save(nuevo);
-
   }
 
-  async updateMensaje(id: number,MensajeActualizar: CreateMensajeDto): Promise<Mensaje> {
+  async updateMensaje(id: number,updateMensajeDto: UpdateMensajeDto,user: any): Promise<Mensaje> {
     const mensajeUpdate = await this.mensajesRepository.findOneBy({ id });
     if(!mensajeUpdate){
       throw new NotFoundException(`No se encontró el mensaje con id ${id}`);
     }
-    mensajeUpdate.nick = MensajeActualizar.nick;
-    mensajeUpdate.mensaje = MensajeActualizar.mensaje;
+
+    if (mensajeUpdate && mensajeUpdate.nick !== user.nick) {
+      throw new ForbiddenException('No tienes permiso para editar este mensaje');
+    }
+    mensajeUpdate.mensaje = updateMensajeDto.mensaje;
     return await this.mensajesRepository.save(mensajeUpdate);
   }
 
@@ -39,7 +39,18 @@ export class MensajesService {
     return this.mensajesRepository.findOneBy({ id });
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number,usuario: any): Promise<void> {
+    const mensaje = await this.mensajesRepository.findOne({
+      where: { id },
+      relations: ['usuario'],
+    });
+
+    if (mensaje && mensaje.nick !== usuario.nick) {
+      throw new ForbiddenException('No tienes permiso para borrar este mensaje');
+    }
+    if (!mensaje) {
+      throw new NotFoundException('Mensaje no encontrado');
+    }
     await this.mensajesRepository.delete(id);
   }
 }
